refactor from async to promise
